move market actions to MarketArea
